% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-mcpf.R
\name{calct1t2}
\alias{calct1t2}
\title{Calculation of Pseudo F Statistic Matrix.}
\usage{
calct1t2(Y, point.seq, H1, H2)
}
\arguments{
\item{Y}{the projection of pseudo outcome data.}

\item{point.seq}{a vector containing the dividing points.}

\item{H1}{the difference between two projection matrixes.}

\item{H2}{the difference between the identity matirx and a projection
matirx.}
}
\value{
a length(point.seq)*2 matrix of pseudo F statistics.
}
\description{
\code{calct1t2} calculates the pseudo F statistics of the divided similarity matrixes
at given dividing point.
}
