% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-mcpf.R
\name{Fibonacci}
\alias{Fibonacci}
\title{Generate Fibonacci Sequence to be Used as Dividing Points}
\usage{
Fibonacci(n)
}
\arguments{
\item{n}{a number indicating the maximum of the generated sequence is
less than n.}
}
\value{
a vector.
}
\description{
\code{Fabonacci} generates a Fibonacci sequence to be used as dividing
 points in MCPF. So it deletes the first element 1 in the original
 Fibonacci sequence.
}
