% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simreg.R
\name{simreg}
\alias{simreg}
\title{Significance Calculation of the Pseudo F Statistic in Distance-based
Regression Model.}
\usage{
simreg(
  simi.mat,
  dis.mat = NULL,
  null.space,
  x.mat,
  permute = TRUE,
  n.monterCarlo = 1000,
  seed = NULL
)
}
\arguments{
\item{simi.mat}{a similarity matrix.}

\item{dis.mat}{a distance matrix. Only one of these two matrixes needs
to be given. If dis.mat is given, it is then transformed to the
similarity matrix.}

\item{null.space}{an index vector containing the index of ${X}_1$,
the covariate matrix to be adjusted for.}

\item{x.mat}{a predictor matrix which contains the covariates and
the predictors of interest.}

\item{permute}{logical.Should sampling with replacement.}

\item{n.monterCarlo}{the number of permutation replicates.}

\item{seed}{the seed for permutation.}
}
\value{
the p value of pseudo F test statistic.
}
\description{
\code{simreg} calculates the p value of the pseudo F statistic in distance
regression model.
}
\examples{
library(MASS)
n = 50
p = 100
k = 15
sigmax = diag(rep(0.5,k)) + matrix(0.5,k,k)
sigmay = diag(rep(1,p))
for(i in 1:p){
  for(j in 1:p){
    sigmay[i,j] = 0.5^abs(i-j)
  }
}
r1 = 0.05
beta0 = r1*matrix(rbinom(k*p,1,0.9), k, p)
x = mvrnorm(n, rep(0,k), sigmax)
y = x\%*\%beta0 + mvrnorm(n, rep(0,p), sigmay)
Ky = calcKerMat_Cen(y)
simreg(Ky,null.space = 1:5,x.mat = x)
}
